backend/requirements.txt
fastapi
uvicorn
sqlalchemy
aiosqlite
httpx
python-dotenv
google-generativeai
pydantic-settings
dnspython
beautifulsoup4
python-whois
